UAEC-17919 set welcome view/account list view

+++ networking/src/main/java/co/yap/networking/leanteach/requestdtos/GetPaymentIntentIdModel.kt
@@ -3,7 +3,7 @@ package co.yap.networking.leanteach.requestdtos
 import com.google.gson.annotations.SerializedName
 
 data class GetPaymentIntentIdModel(
-    @SerializedName("amount") var amount: Int? = null,
+    @SerializedName("amount") var amount: Double? = null,
     @SerializedName("currency") var currency: String? = null,
     @SerializedName("customer_id") var customerId: String? = null
 )
